split the schemas
